+++ src/components/equipment/helmet.rs
@@ -42,12 +42,12 @@ impl Helmet {
         for _ in 0..rarity.n_affix() {
             match provider.gen(rng) {
                 Some(HelmetAffixKind::AddArmour) => {
-                    let affix = Armour(rng.gen_range(1. ..=3.));
+                    let affix = Armour(rng.gen_range(1..=3) as f32);
                     labels.push(affix.to_string());
                     helmet_commands.insert(affix);
                 }
                 Some(HelmetAffixKind::AddLife) => {
-                    let affix = MoreLife(rng.gen_range(5. ..=10.));
+                    let affix = MoreLife(rng.gen_range(5..=10) as f32);
                     labels.push(affix.to_string());
                     helmet_commands.insert(affix);
                 }
